projsort/projsort/requirements.txt
pdfminer.six==20240706
Pillow==10.4.0
pandas==2.2.2
Jinja2==3.1.4
tqdm==4.66.5
Pygments==2.18.0
matplotlib==3.9.0
PyYAML==6.0.2
